refactor: flatten message fetcher module
